WIP: port verity 5.7.3 (1.20.1 Forge) to 1.21.1 NeoForge
- Decompile 5.7.3 and copy sources into project
- SRG m_/f_ remap to 1.21.1 official names (296/450 methods, 170/170 fields)
- Forge->NeoForge package remap (net.minecraftforge.* -> net.neoforged.*)
- Add YACL dep; conditional sherpa-onnx libs dep
- Sync remote bugfix 508bff5 AiAPI quality (SLF4J, shared HttpClient)

+++ src/main/java/varmite/verity/entity/client/VerityAnimation.java
@@ -1,8 +1,14 @@
 /*
  * Decompiled with CFR 0.152.
+ * 
+ * Could not load the following classes:
+ *  varmite.verity.entity.client.VerityAnimation
  */
 package varmite.verity.entity.client;
 
+/*
+ * Exception performing whole class analysis ignored.
+ */
 public class VerityAnimation {
     public static float getYOffset(float tick) {
         if (tick < 0.0f || tick > 50.0f) {
@@ -47,7 +53,7 @@ public class VerityAnimation {
     }
 
     public static float getScaleXZ(float tick) {
-        float scaleY = VerityAnimation.getScaleY(tick);
+        float scaleY = VerityAnimation.getScaleY((float)tick);
         return 1.0f + (1.0f - scaleY) * 0.75f;
     }
 }
